Downgrade payload to 3.6 due to join field bug, migrate to usages field where it makes sense and align collections code structure
