+++ cms/src/collections/new-pages/index.ts
@@ -1,0 +1,1 @@
+export * from "./new-pages";
