+++ cms/src/collections/texts/texts.ts
