+++ cms/src/fields/rich-text.ts
@@ -1,5 +1,4 @@
 import { RelationshipField } from "payload";
-import { Texts } from "../collections/texts";
 
 export type RichTextFieldOptions = {
   optional?: boolean;
